Fix corruption issue in fj_are_structure affecting PST
The function tried to treat projectiles in PST areas, which do not
have projectiles. It has been fixed to only handle projectiles if the
game is not PST.

The documentation has been amended to state that projectiles are not
available for PST and if you try it anyway, the mod fails with an
uninformative error. This could be handled better, but the function
does not presently have any sanity checking of input (which is itself
something of an issue).