Add the root game path to the default list used for locating biffs
Biffed files are typically stored with the data/ part of the path
included in the file name.

The default list is not considered at all, or only searched after all
paths read from INI files.

In conventional game setups, this should eliminate the need for a
linux.ini, or equivalent